`Communication`: Allow saving posts for later (#307)
* Add save post function to message service

* Split MessageActions

* Refactor Edit/Delete into MVVM

* Move pinning to ViewModel

* Remove unused variable

* Refactor edit message

* Change to Int64

* Add button to add/remove bookmark

* Add localizations

* Add indicators for saved messages

* Use intermediate core modules version

* Use newer commit
