+++ ArtemisKit/Sources/Messages/Views/MessageDetailView/MessageCell.swift
@@ -29,6 +29,7 @@ struct MessageCell: View {
             reactionMenuIfAvailable
 
             VStack(alignment: .leading, spacing: .s) {
+                savedIndicator
                 pinnedIndicator
                 resolvesPostIndicator
                 headerIfVisible
@@ -118,6 +119,10 @@ private extension MessageCell {
         (message.value as? Message)?.title ?? ""
     }
 
+    var isSaved: Bool {
+        message.value?.isBookmarked ?? false
+    }
+
     var isPinned: Bool {
         (message.value as? Message)?.displayPriority == .pinned
     }
@@ -149,6 +154,7 @@ private extension MessageCell {
 
     var messageBackground: Color {
         useFullWidth ? .clear :
+        isSaved ? .blue.opacity(0.2) :
         isPinned ? .orange.opacity(0.25) :
         resolvesPost ? .green.opacity(0.2) :
         Color(uiColor: .secondarySystemBackground)
@@ -163,6 +169,13 @@ private extension MessageCell {
                 verticalPadding: .s
             )
             .font(.footnote)
+        }
+    }
+
+    @ViewBuilder var savedIndicator: some View {
+        if isSaved {
+            Label(R.string.localizable.savedMessage(), systemImage: "bookmark")
+                .font(.caption)
         }
     }
